Add platform specific extension to TestContext
To facilitate storing platform specific elements in TestContext, a new
PlatformTestContext expect/actual has been introduced. This can be
accessed from the TestContext with `testContext.platform`. When using
this on a certain platform, you will have access to all the platform
specific elements defined for that platform. Currently, no elements have
been added for any platform, but in a future commit the Android version
will get an AccessibilityValidator.

Bug: 198122348
Test: presubmit

Change-Id: Iddc3d2585267a181161d5774df6b658f73155462