Avoid a deadlock in the Recomposer accessing the frame clock
There is a potiential deadlock when in the Recomposer when it checks
the frame clock for awaiters as it aquires the frame clock lock. These
locks are normally taken in the opposite order.

This change allows the Recomposer to check for awaiters without
taking the frame clock lock, avoiding the deadlock.

Fixes: 329011032
Test: ./gradlew :compose:r:r:tDUT
Change-Id: Iff3f93b195b92adcc58210ae4a7f633bac390bb3