Fixes ripples added before the ripple node was drawn being invisible
Adding a ripple before the node was drawn causes the target radius to be the default value of 0. Radius updates for existing ripples aren't supported by RippleDrawable, so this means that the ripple would be 'invisible'. Instead, now we delay adding a ripple until we know the size, and we do this before draw so that we can draw the ripple correctly on the first frame, instead of needing to wait another frame to draw the ripple.

Bug: b/326274138
Fixes: b/329693006
Test: RippleModifierNodeTest
Change-Id: I6b99a950d5ac9c71b6b6995560473fde321d9ccd