Fix JNI bundling of SQLite KMP in AAR artifact.
The AndroidXMultiplatformExtension's addNativeLibrariesToJniLibs() API was using the 'debug' variant as default variant to add JNI files to, however artifacts are built from the 'release' variant, while tests use the 'debug' one. The API has been updated to add the JNI all variants.

Also remove `androidNative()` as a KMP target of sqlite projects since even though possible, we don't expect the library to be used in an Android native project.

Test: Validated sqlite-bundled-android-2.5.0-alpha01.aar contained jni directory with native files.
Change-Id: Iff8351ea52979026d28cedb525db835d491839ac