Update anchoredDrag behavior
This change updates anchoredDrag's target-less overload to not snap to the closest anchor at the end of an anchoredDrag anymore, as this was impacting `drag` functionality. This contract is not maintainable at the draggable integration level as flings happen in a different scope than the user input anchored drag interaction.

Test: anchoredDraggable_customDrag_doesNotSnapToClosestAnchor
Relnote: Reverted a recent contract change where AnchoredDraggableState#anchoredDrag calls would snap at the end of a anchoredDrag operation.
Change-Id: I9571542e53b0bd026e6bcb8ca6abead58ca4a4ae