Remove unnecessary key event routing
Before aosp/2813125 and aosp/2901550, when we cleared focus in Compose,
we would not clear focus from the host AndroidComposeView. This led to a
situation where AndroidComposeView would receive keyEvents when no
focusTargets were focused. In this case we would find the first onKeyEvent
modifier before the root focus node and route the key event to it.

We don't have this situation anymore since we clear focus from
AndroidComposeView whenever we clear focus in Compose. This CL removes
the code which would try to find the first onKeyEvent modifier even if
there was nothing focused in Compose. Since we don't consume the keyevent
anymore, the framework will route this to the currently focused view.

From the stack trace in the attached bug, it looks like this logic used to
cause a crash when we attempt to route a key event when we don't
have an active target. The exact use case which triggered this crash
isn't clear, but removing this logic will also fix this bug.

Bug: 325709077
Test: All existing tests pass
Change-Id: I05d842d33d4383b50027f44509fc6bec10be1c30