Add NavController DSL to create graph with safe args
Added two NavController.createGraph overloads to create graphs with startDestination from a serialiazable KClass or from a serializable Object.

To support object-based startDest, we need to add a null check while parsing startDestination args. Previously a `isDefaultValuePresent == true` sufficiently indicates the existence of a non-null defaultValue for non-nullable types. However in the case of Safe Args, it is possible for `isDefaultValuePresent == true` while defaultValue is not populated (meaning it is still null). This is due to serialization limitations where we know a defaultValue is present but cannot read actual value.

Test: ./gradlew navigation:navigation-runtime:cC
Bug: 188693139
Relnote: "Added NavController createGraph DSL to create graphs with startDestinations from either a serializable KClass or a serializable Object. Argument values from the serializable Object will be parsed automatically and available in the NavBackStackEntry's argument bundle."
Change-Id: I836a1559471fca7594653ea9e4e6a53b0d1b7ff0