Remove super.close() calls during SQLite open helper recovery steps.
There is no need to close() the database as internally if an error occurs during opening an a database was created but not yet assigned to be returned, then it is automatically closed by the framework. Having the close() calls during recovery can swallow exceptions if the close() fails due to a recursive open.

Bug: 329311309
Test: OpenHelperRecoveryTest
Change-Id: I2f71def02e29ac3fdab46aae08d572b3dfa19ad9