Disabling Task Based UX in Studio Profiler
Android Studio Profiler switches to Task Based UX from the Koala release
(first in Canary 2). This conflicts with the flag that redirects
Perfetto traces to ui.perfetto.dev.

As an intial mitigation, we're disabling the task based UX (ahead of
Studio updating to Koala), while working on support for ui.perfetto.dev.

Bug: 331398285
Test: tested relevant flag args on Android Studio Koala Canary 2
Change-Id: I9366a7c27b71873cf19fcba14833c5b76ca400c5