Create AndroidFragment to allow adding fragments in Compose
Adding new AndroidFragment API that creates the given Fragment class and
adds it to the enclosing FragmentManager, within Compose.

It handles setting the ViewCompositionStrategy and saving the state of
the fragment across recompositions.

RelNote: "The new AndroidFragment Composable allows adding fragments
into the Compose hierarchy via the fragment class name."
Test: added AndroidFragmentTest
Bug: 230126482
Bug: 312895363

Change-Id: Icf84199bbe487b2a2b6a95d2b6e09415f810e77a