Implement links interactiom listeners and styling through annotation only
LinkInteractionListener is set on LinkAnnotation. Currently only has a click function but is a fun interface. In the future can be extended to support onHover or other interactions that are needed. Note that Clickable should ideally have it non-null but this wouldn't work as we need to restore it.

LinkAnnotation now takes the style, focusStyle and hoverStyle.

Test: ui:ui-text:test, ui:ui-text:cC, f:f:cC in text folder
Relnote: "LinkAnnotation now takes the state-based styling arguments and a LinkInteractionListener. Add this annotation to the AnnotatedString to get a hyperlink. By passing focusedState and/or hoveredState you can define the visual conficuration for links when they are focused and/or hovered."
Bug: 139312671

Change-Id: I81ce4350b8a1e37881000fd82f081b7afb8e0f42