API change: Consolidation ImageAnalysis coordinate system constants
Move all constants, original, view and sensor to ImageAnalysis. Deprecate the view constant in CameraController. This is for addressing API council reiview.

Relnote: Replace CameraController#COORDINATE_SYSTEM_VIEW_REFERENCED with ImageAnalysis#COORDINATE_SYSTEM_VIEW_REFERENCED. The value of the constant remains the same. This is for consolidating all the constants into one place.

Bug: 331480446
Test: manual test and ./gradlew bOS
Change-Id: I890bb13e2c045393bd7b6e438c0f35a03eefe240