Wrapping `checkIdentity()` in transaction.
This wrapping is meant to guard against the potential scenario where the
call to `updateIdentity()` in `RoomConnectionManager` is interrupted
between the creation of a master table and creation of the insert
query that inserts the identity hash to the room_master_table.

In the scenario where there is an interrupt, the master table will be
created but will have a null hash as well as no rows. This will
cause `checkIdentity()` to fail when the app is reopened, causing
an `IllegalStateException`.

Bug: 322668000
Test: OpenExistingDbTest.kt
Change-Id: Ia02a033addba4b4794825d45ab7e348f0d5b3c73