[glance] Add padding override to Scaffold and test
Allow padding to be overridden. Allow TitleBar to be null. Add a screenshot test
Update the default horizontal margins of scaffold to be 12dp
Update scaffold to use the system defined corner radius dimen
Remove 4dp of padding from TitleBar so that its icon aligns with
the 12dp of scaffold padding.

Test: Screenshot tests
Relnote: Adds optional parameter to Scaffold

Change-Id: If753f541b74ca0e2bb532ca2fd61ffe3ce240400