Fixes overscroll consuming deltas and velocity incorrectly in pre scroll / pre fling
Previously overscroll was stretching and relaxing in pre scroll, and stretching in pre fling. The actual intended / correct behavior here is that we only relax in pre scroll and pre fling, and stretching should only happen in post scroll and post fling. This means that nested scroll connections can correctly see available delta / velocity before overscroll increases the stretch, whereas before they didn't see this properly.

This CL fixes these issues, but needs to be landed with some other fixes, as currently overscroll consumes all delta when relaxing, which means that flings will not be propagated with remaining velocity, so the list will just not move if you fling to relax. (this was not visible before as we incorrectly relaxed post-fling)

Fixes: b/310186412
Fixes: b/288499761
Test: StretchOverscrollIntegrationTest
Relnote: "Fixed an issue where overscroll was incorrectly consuming some events before NestedScrollConnections could see the delta / velocity. Now overscroll will only consume before to relax an existing stretch."
Change-Id: I147b3491a4fbb21a4b83734ba8c22b4d549f888d