Add a in-process and multi-process lock during database configuration.
An exclusive lock that protects a region from concurrent in-process threads and other processes is used during first time database initialization, i.e. when the first database connection is opened and the database has to validate schema, maybe perform migrations or first time schema creation. Note that multi-process locking is only needed for file databases (not in memory) and only during that first time initialization.

Also guard against recursive initialization and usage of the special name ':memory:' as a file database name.

Bug: 316945717
Test: FileLockTest + BuilderTest
Change-Id: Ic87280ce534ca235a4600cb059c6bac6147fccac