Add DynamicNavGraphBuilder safe args constructors
Add support to constrcut a DynamicNavGraph with route from a serializable KClass and startDestination from either an Object or a KClass.

Also added #navigation DSL to create root and nested DynamicNavGraphs.

Test: ./gradlew navigation:navigation-dynamic-features-runtime:cC
Bug: 188693139
Relnote: "Add support to constrcut a DynamicNavGraph with route from a serializable KClass and startDestination from either an Object or a KClass"
Change-Id: Ia668dffcae8b255bf5db75b6c149983d973d2b06