Fetch ui for ongoing session
Fetches UI for a session when surfaceView is attached to window.
This lets the previously presented provider UI remain visible when
it is scrolled back into a view. Also adds a manual test
and an Integration Test to verify that the solution
works as expected.

Bug: 333442774
Test: testPoolingContainerListener_NotifyFetchUiForSession && UiLib Manual Test
Change-Id: I804dfd4ab8736593ed310df2f72e5ae37b73b381
Relnote: "Fixed rendering bug when displaying remote content inside pooling containers."