Adds APIs to FocusTargetModifierNode for observing focus changes, configuring focusability, and requesting focus
This allows delegating nodes to just delegate to one FocusTargetModifierNode, instead of needing to add separate requester / event / properties nodes to modify the delegated and owned target node.

Also migrates focusable / clickable to use this, which reduces the cost of focus invalidation as we can remove all the other node types.

FocusProperties APIs for configurating focus requesters will be added in the future to FocusTargetModifierNode.

Bug: b/298048146
Test: FocusableTest
Test: ClickableTest
Test: FocusTargetModifierNodeTest
Relnote: "Added FocusTargetModifierNode#requestFocus(), FocusTargetModifierNode#focusability, and a new FocusTargetModifierNode() overload that exposes FocusTargetMode and a callback for when focus state changes. This makes it simpler to implement delegated modifiers that should be focusable, as you do not need to implement requester / event / properties nodes as well, in order to modify the target node."
Change-Id: I27f84d34307dc78bbd14ecda328ea9cdc5b63b8b