Fixes issue where modifier node updates were happening while unattached
In rare cases, such as when using movable content inside a subcomposition, it's possible for the modifier to be updated while the node is unattached. This means that update code that relies on being attached (such as to invalidate semantics) will cause crashes. Instead we should guard updating the modifier until when we are attached.

Bug: b/327814644
Fixes: b/332814226
Test: ClickableTest
Test: CombinedClickableTest
Test: LayoutNodeTest
Change-Id: I36e9ce509529017315b14267b269c4aeeff1ef18