ExifInterfaceTest: Clean up naming of GPS expected values
The gpsXXX values are the ones actually stored in the file, and their
names correspond to the Exif attributes. The floating point expected
longitude, latitude and altitude values relate to the values computed
by ExifInterface and returned from getLatLong() and getAltitude().

This change aims to tidy up the naming of these expected values, and the
expected offset and length values, to make the distinction a bit
clearer.

This change also removes some TODOs about introducing arrays in a couple
of places. I decided this was a bad idea, because of the inherent
mutability of java arrays. I considered ImmutableList instead, but that
doesn't really simplify the assertions much beyond what they are now.

Bug: 270554381
Test: ExifInterfaceTest
Change-Id: I67edc17e94542b9bbf8f08cbf0c07b0ad30c378c