Fix KType mismatch with kotlin reflect
When mapping an argument Type to NavType, the matching used to be based on type names. However the type name varies depending on whether build dependency has kotlin reflect.

Refactorted implementation to be based on hashcode which is generated at runtime. This means the matching is independent of kotlin reflect and provides more reliable matching.

Test: ./gradlew navigatiion:navigation-common:test
Bug: 333965952
Bug: 188693139
Change-Id: Ia660ee2880558f19af58cd6fbe240946374eb7c4