New add deepLink from KClass APIs for NavDestinationBuilder
NavDestinationBuilder supports adding deepLink from a reified serializable KClass by taking a base path and then appending arguments extracted from the KClass.

The deepLink does not take a typeMap and instead relies on the arguments that were added when initializing NavDestinationBuilder with route from KClass. As such, the deeplink is expected to have identital arguments (in name and type) as the destination KClass.

Test: ./gradlew navigation:navigation-common:cC
Bug: 188693139
Relnote: "Add NavDestinationBuilder deepLink APIs to add DeepLinks based on serializable KClass"
Change-Id: I3a64b21c11307c1cb409e2bb47e292021d55c0b7