Fix per-connection configuration in Room's connection manager
In aosp/3028503 logic was added to perform initial database configuration once in an exclusive lock, but it broke per-connection configurations since once the 'isConfigured' flag was flipped no other configuration was done on opened connections. This CL fixes the manager to still perform some non-initial database configuration on connections being opened for the pool.

Bug: 335244484
Test: BuilderTest
Change-Id: I0efc54e87dc8d939398481468c4f94ce610c9dfc