[KSP2] Fix star projections
`KSTypeArgument.type` of star projections are no longer resolved to upper bound and have a `null` value in KSP2 and this leads to NPE in XProcessing. In this CL we try to accommodate that by introducing a new type `KspStarTypeArgumentType` to represent star projections.

Unlike in KSP where `KSTypeArgument` is not a `KSType`, in XProcessing `KspTypeArgumentType` is a `KspType` so we still need a `KSType` in `KspStarTypeArgumentType`. Here we follow `KspVoidType` and use `Unit` as a placeholder.

The type names we generate will still be `?` in Java and `*` in Kotlin. This is the same for both KSP1 and KSP2.

Bug: 314151707
Test: XTypeTest.kt, TypeAssignmentTest.kt
Change-Id: Idb6fc262f342a73a2f524230ae3999102c2889a0