Enable GraphicsLayerOwnerLayer on API levels M..O
On those Android versions we use GraphicsLayerV23. We will enable it on versions using GraphicsViewLayer separately.
One of tests caught an issue where we were discarding the layer content before the persistence logic had a change to be executed. In order to fix it I updated the handler scheduling to use sendMessageAtFrontOfQueue.

Test: existing test from all the compose modules
Change-Id: Ic61a639009b1759688b304ae5d25f040954042b6