Add hasRoute<T> API for NavDestination
Convenience method to check if the NavDestination's route is based on KClass<T>.

Internally, the implementation is based on destination ID which was set during its initialization. Therefore if the ID was later overridden with a custom ID, the function will always return false.

Test: ./gradlew navigation:navigation-common:cC
Bug: 188693139
Relnote: "Added hasRoute<T> API for NavDestination to check if the NavDestination's route is based on KClass<T>."
Change-Id: I6c3a2109d99a3c40102b8a81d7a8bc17746c6a93