Disable Java-side support for ReplaceWith on Kotlin properties
We don't currently map property expressions to/from their corresponding
accessor methods, so we should avoid showing replacements for properties.

Note that we're still showing replacements on physical functions where the
replacement expression is a property -- that's hard to ignore, since we'd
need to parse the expression to know that it's a property.

Bug: 323214452
Test: ReplaceWithDetectorTest
Change-Id: I5dd76d71969199535da5d38daad622a082b0ee69