Fix the initial state issue of SizeTracker
We missed to apply the same logic to address the initial state problem
in PositionTracker to SizeTracker as well.

More details: during the first time Modifier.animateBounds is invoked,
the current size will be IntSize.Zero, i.e., the initial (invalid) value
of size to animate. If the target size keeps the same afterwards,
the original size will never get a chance to be updated and remain as
IntSize.Zero, which causes incorrect size animations. We apply the same
logic as PositionTracker here: when the current size is invalid, just
use the new target size as the original size as in this case no
animation is needed, so the original size can be safely set to the
target size.

Note this doesn't fully solve the case that when the current animation
is interrupted by anohter animation with the same target size. But it's
a minor case and requires more sophiscated logic to address. We will fix
that later.

Test: TODO, we need to figure out how to test animation properly
Bug: 335687032
Change-Id: I35cb1b6a702aec3dd940bc4eacce0c0e5a64f22a