Fingerprint should include the the absolute position of children inside the parent
Previously the childNodeValue was creating a fingerprint which was unique enough at a single level.
But it was ignoring the number of grand children for each added node.
In a layout like [a b][c d] this would end up being 31^2*a+31*b+31*c+d (assuming the parent containers are similar).
With that fingerprint, b and c can change place and the fingerprint won't change.

With this change we now consider how many subnodes (children, grandchildren, ...) is also included in each node addition.
The new fingerprint will be 31^4a+31^3+31*c+d

Bug: 332257203
Change-Id: I3f235fdc7ae0cbb21d3e98f4dc0b776b1c2d18f8