Improve ContextMenuArea first compose performance
This pushes the string resource reads of the context menu item labels to when the context menu items are composed instead of when the ContextMenuArea (currently wrapping `SelectionContainer`s and `BasicTextField`s) is composed. This results in the time taken to resolve labels to be incurred when the context menu opens, and not immediately when the `SelectionContainer`/`BasicTextField` is composed for the first time, which is a much hotter path.

Test: Existing tests
Fixes: b/335633648
Change-Id: If443b6e9f0593b320317ba1646a6067dc2769478