Add SessionObserver interface and registration logic
Introduces a SessionObserver interface which is used as a
way to receive updates about the presentation state of a UI
session. This interface will receive updates when a session
is opened, when its UI presentation changes, and when the UI
session is closed.

As one or more UI sessions may be created for a SandboxedUiAdapter,
we introduce the notion of a SessionObserverFactory. This factory
may be registered to a SandboxedUiAdapter. For every UI session opened
for an adapter, a new SessionObserver will be created.

Relnote: Add SessionObserver interface and registration logic
Test: IntegrationTests
Bug: 309940664
Change-Id: I047dc591b385780fb8c8b914832165b48ca8727d