Add tests for ExifInterface 'XMP from separate JPEG marker' handling
These tests enforce the documented behaviour for JPEG that, if both
are present, XMP in the Exif segment is preferred. This is contrary
to the XMP spec which says:
> The TIFF in the Exif APP1 marker segment should not contain tag 700
> (XMP)

Unfortunately changing this behaviour to align with the spec could
introduce unexpected behaviour for apps relying on the previous
behaviour.

This change also adds some robustness to the implementation so it no
longer assumes the Exif APP1 segment comes before the XMP one.

Test: ExifInterfaceTest
Change-Id: I1a8d99649d699671fbdca93ef2e9fa6248804b13