Fix JVM conflict with getRequiredTypeConverterClasses()
When the @Database annotated class is in Java and being processed in KSP with Kotlin Codegen, the compiler would not consider `getRequiredTypeConverterClasses()` as a function to override because it conflicted with the internal property. This CL fixes the issue by simply renaming the internal property.

Moreover, generate a start type projection to match the actual function being overridden.

Bug: 337300565
Test: DatabaseKotlinCodeGenTest
Change-Id: I5f79aedf8489ed4886b585cd2ac6c249f461d7a1