switch to speaker in CallSessions; inc to alpha05
This CL fixes a recent regression caused by using Channels#receive().
The receive API caused the flow to only be consumed in the CallsManager
layer and not echo the values to the client.

Forward fixing the approach, the CallSessions now cache the endpoint
info and make a one time attempt to switch at the start of the call.

Fixes: 336788058
Test: manual +
      CallControlScopeFlowsTest & CallSessionTest &
      testAddOutgoingVideoCall_CallEndpointShouldBeSpeaker*
Change-Id: I4f45f8aa6bf882373383feb1be5205896759ed4a