Extract interface from SharedTransitionScope
SharedTransitionScope should match other scope interfaces and allow for
implementation. It also doesn't need to publicly expose its
`coroutineScope` property. This changes `SharedTransitionScope` to be an
interface, implemented by `SharedTransitionScopeImpl`.
Its public functions and types are in `SharedTransitionScope`, and its
implementation functions and helpers are in `SharedTransitionScopeImpl`.

Note: The internal `sharedBoundsWithCallerManagedVisibility` is in
`SharedTransitionScopeImpl`, so using it requires a cast. I personally
believe this should be made public, as I've already found use cases for
it, but that doesn't need to be done here.

Relnote: Experimental SharedTransitionScope is now an interface rather than a class
Test: existing tests pass

Fixes: 338415048
Fixes: 338414702

Change-Id: Iaf856b84ad2d91f94f8e294f015b6341808fcc74