[Compose] Cleanup and API feedback for ConstraintLayout
Applying API feedback changes to ConstraintLayout and MotionLayout.

Some minor code cleaup and a couple of tests fixed.

Removed blank lines in documentation (nbsp;) as the renderer in Studio and DAC have changed so that additional blank lines are not necessary.

Relnote: "Simplified `animateChanges` API in ConstraintLayout to only need a non-null AnimationSpec. `onIncomingConstraints` in InvalidationStrategy is now a regular lambda. `fixedWidthRate` renamed to `shouldInvalidateOnFixedWith`, same for the Height variant."

Bug: 291769109
Bug: 298404266
Bug: 332898040
Bug: 336370035
Test: N/A
Change-Id: Ie59cd551b999788c0f11ce77feb62d00dc8b9de6