Fix modified to no longer change the composite hash code
The composite hash code is now calculated differently
and the implementation of `materialize` calls the runtime
in a way that always differed from the way the compiler
would have generated its code, which now calls the runtime
in a way that it doesn't support.

With this change, materialize will now unconditionally
generate a group for the `ComposeNode` that calls it so
that the composite hash code will no longer change if a
materilized modifier switchs from generating and not
generating a group.

Fixes: 338478720
Test: new test, ./gradlew :compose:r:i-t:cC
Change-Id: I5b17576fafd924cfeb17066f9b7ab3d0bc4a19ca