Ensure that all fragments are seekable to run preditive back
Currently, if you have a Animator or Seekable Androidx Transition on any
fragment it will always run regardless of it the other fragments in the
transaction are also seekable. We need to ensure that it is always all
or nothing, either all fragments in the transaction are seekable and we
do predictive back or one of them is not seekable and we hold the
transition.

RelNote: "Predictive back will now only run for transactions in which
all of the fragments have either a Animator or a Seekable Androidx
Transition. This fixes an issue where cancelling a partially seekable
transaction would cause a black screen."
Test: Added FragmentTransitionSeekingTest
Bug: 339169168

Change-Id: I4303795b1cb4f7e58cd4f8efe0baa53d64a51dd8