Treat platform types as nullable by default
This causes KotlinPoet names to be nullable when the type is a platform type, i.e. from Java sources or compiled Java code. Some exceptions are added for types that are never null, these are: annotation value types, exception types in 'throws' declarations, super types and in Kotlin only the array type of a vararg.

Two additional changes that contribute to this CL are:
 * Remove workaround for https://github.com/google/ksp/issues/1004, it is no longer needed and the KSType for annotation values are correctly returned by the latest version of KSP along with the correct nullability.
 * Remove 3 lazy properties in KspProcessingEnv, these are the KSTypeVarianceResolver, KspArrayType.Factory and CommonTypes. All 3 helpers retained KSP's resolver across rounds, which shouldn't be done as it can lead to stale caches.
 * Avoid usages of toKS() to check for KSP 2, instead create a helper internal function.

See also https://kotlinlang.org/docs/java-interop.html#notation-for-platform-types

Bug: 331427836
Test: ./gradlew :room:room-compiler-processing:test
Change-Id: I643ba1d9c005b1b5c5084d08a23288298bec7489