AndroidX Biometric Codes over Framework Codes
The decision was made to convert the framework codes that arrive to
AndroidX into AndroidX specific codes from the Biometric library.
Namely, this change converts all the received errors, to their jetpack
counterparts. This repeats this for the Allowed Authenticators as well.

The values should be identical, and some values are still not included
in the framework, but are included in JP, such as the 'negative button'.
So this should continue to be modified over time.

Similar changes can follow for the authentication result.

Bug: 340334264
Test: JP Tests included (unit)
Change-Id: I047566912f976612aee8040940a9b18a8d36a7ba