Separate pooling container tests from other integration tests
Separates tests relating to pooling containers to their own class, and
abstracts common logic to utility classes: one for session management
and one for screenshot testing.. This is done to avoid
IntegrationTests becoming too large, and to avoid the different
test views interfering in the MainActivity.

In the future, this can be improved by providing a base class that
each test class can subclass.

Test: IntegrationTests, PoolingContainerTests
Bug: 339384188
Change-Id: Iff1500231cca17c48cf87eeccb900918e4b47937