Add a parameter to curvedComposable to avoid composables rotation.
This is useful when a developer wants to lay out compnents on an arc but doesn't want does components rotated.
Note that we are not changing the computations of the size of the item since that will introduce a circular dependency between size and position.

Relnote: "Added parameter to CurvedLayout.curvedComposable to avoid components being rotated."
Test: Manual

Change-Id: I668986bc70f5d46873c43e0ca9755ab71fd296b6