Disabling opening traces in Perfetto Web UI
Temporarily disabling opening traces in Perfetto Web UI. Perfetto
Web UI changed the way they're fetching local traces. We fixed it in
ag/27303635, but we need to wait to upgrade to a Studio build that
contains the fix.

The fix also addresses the new Studio UI (task based flow), so we no
longer need to disable that (removed the entry from the config file).

Bug: 331398285
Test: manual
Change-Id: I8d6f8f9c74fc856b134162bfec59e2e6864bb257